202603121722 change uint8 to uint64

+++ gmm/grouped_matmul/op_host/op_tiling/grouped_matmul_tiling.h
@@ -50,7 +50,7 @@ TILING_DATA_FIELD_DEF(uint64_t, quantGroupNum);        // for A8W4 MSD
 TILING_DATA_FIELD_DEF(uint64_t, isPreTiling);
 TILING_DATA_FIELD_DEF(uint32_t, withOffset);
 TILING_DATA_FIELD_DEF(uint32_t, isOutputDisableL2Cache);
-TILING_DATA_FIELD_DEF(uint8_t, isA4W4Optimize);
+TILING_DATA_FIELD_DEF(uint64_t, isA4W4Optimize);
 END_TILING_DATA_DEF;
 REGISTER_TILING_DATA_CLASS(GMMBaseParamsOp, GMMBaseParams)
